+++ app/src/main/java/com/hyntix/android/usbflasher/ui/theme/Theme.kt
@@ -1,6 +1,6 @@
 package com.hyntix.android.usbflasher.ui.theme
 
-import android.app.Activity
+
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme